Work in progress to remove cuncurrency problems.
