+++ faostat-api-core/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
     <groupId>org.fao.faostat</groupId>
     <artifactId>faostat-api-core</artifactId>
-    <version>1.2.0-SNAPSHOT</version>
+    <version>1.2.1-SNAPSHOT</version>
     <packaging>jar</packaging>
     <name>FAOSTAT Core</name>
     <properties>
