+++ faostat-api-web/src/main/java/org/fao/faostat/rest/V10.java
@@ -9,37 +9,4 @@ import org.fao.faostat.core.FAOSTATAPICore;
  * */
 public class V10 {
 
-    @InjectParam
-    DefaultOptionsBean o;
-
-    @InjectParam
-    FAOSTATAPICore faostatapiCore;
-
-    public V10() {
-        this.setO(new DefaultOptionsBean());
-    }
-
-    public void storeUserOptions(String datasource, String apiKey, String clientKey, String outputType) {
-        this.getO().setDatasource(datasource != null ? datasource : this.getO().getDatasource());
-        this.getO().setApiKey(apiKey != null ? apiKey : this.getO().getApiKey());
-        this.getO().setClientKey(clientKey != null ? clientKey : this.getO().getClientKey());
-        this.getO().setOutputType(outputType != null ? outputType : this.getO().getOutputType());
-    }
-
-    public FAOSTATAPICore getFaostatapiCore() {
-        return faostatapiCore;
-    }
-
-    public void setFaostatapiCore(FAOSTATAPICore faostatapiCore) {
-        this.faostatapiCore = faostatapiCore;
-    }
-
-    public DefaultOptionsBean getO() {
-        return o;
-    }
-
-    public void setO(DefaultOptionsBean o) {
-        this.o = o;
-    }
-
 }
