+++ faostat-api-web/src/main/java/org/fao/faostat/rest/V10Glossary.java
@@ -1,5 +1,10 @@
 package org.fao.faostat.rest;
 
+import com.sun.jersey.api.core.InjectParam;
+import org.fao.faostat.beans.DatasourceBean;
+import org.fao.faostat.beans.DefaultOptionsBean;
+import org.fao.faostat.core.FAOSTATAPICore;
+import org.fao.faostat.core.StreamBuilder;
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.*;
@@ -13,10 +18,15 @@ import javax.ws.rs.core.StreamingOutput;
 @Component
 @Path("/v1.0/{lang}/glossary/")
 @Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
-public class V10Glossary extends V10 {
+public class V10Glossary {
+
+    @InjectParam
+    FAOSTATAPICore faostatapiCore;
+
+    private DefaultOptionsBean o;
 
     public V10Glossary() {
-        super();
+        this.setO(new DefaultOptionsBean());
     }
 
     @GET
@@ -28,7 +38,7 @@ public class V10Glossary extends V10 {
 
 
         /* Store user preferences. */
-        this.storeUserOptions(datasource, api_key, client_key, output_type);
+        this.getO().storeUserOptions(datasource, api_key, client_key, output_type);
 
         /* Store procedure parameters. */
         this.getO().addParameter("lang", this.faostatapiCore.iso2faostat(lang));
@@ -36,8 +46,14 @@ public class V10Glossary extends V10 {
         /* Query the DB and return the results. */
         try {
 
+            /* Stream builder. */
+            StreamBuilder sb = new StreamBuilder();
+
+            /* Datasource bean. */
+            DatasourceBean datasourceBean = this.faostatapiCore.getDatasourcePool().getDatasource(this.getO().getDatasource().toUpperCase());
+
             /* Query the DB and create an output stream. */
-            StreamingOutput stream = this.getFaostatapiCore().createOutputStream("glossary", this.getO());
+            StreamingOutput stream = sb.createOutputStream("glossary", datasourceBean, this.getO());
 
             /* Stream result */
             return Response.status(200).entity(stream).build();
@@ -46,6 +62,14 @@ public class V10Glossary extends V10 {
             return Response.status(500).entity(e.getMessage()).build();
         }
 
+    }
+
+    public void setO(DefaultOptionsBean o) {
+        this.o = o;
+    }
+
+    public DefaultOptionsBean getO() {
+        return o;
     }
 
 }
