+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>org.fao.faostat</groupId>
     <artifactId>faostat-api</artifactId>
-    <version>1.2.0-SNAPSHOT</version>
+    <version>1.2.1-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <name>FAOSTAT API</name>
